Replace erroneous type signature
It is not surprising that this not yield any compile-time errors, given the current structure of ZoomCache, but it is a bit disconcerting that it did not yield any runtime errors.

+++ tools/zoom-cache-gnuplot.hs
@@ -41,6 +41,6 @@ main = do
               streams <- getStreams fp tn :: IO [Stream Int]
               plotSummaries lvl streams
           Just ZDouble -> do
-              streams <- getStreams fp tn :: IO [Stream Int]
+              streams <- getStreams fp tn :: IO [Stream Double]
               plotSummaries lvl streams
           Nothing -> return ()
